cleanup some imports
