+++ package.json
@@ -22,7 +22,7 @@
 		"cheerio": "^1.0.0-rc.3",
 		"chokidar": "^3.5.1",
 		"discord.js": "discordjs/discord.js#fe7df708e44e0280dfaf0f8e457b154781bb5140",
-		"e": "^0.1.3",
+		"e": "^0.2.0",
 		"emoji-regex": "^9.2.1",
 		"fastify": "^3.11.0",
 		"fastify-cors": "^5.2.0",
