+++ src/lib/data/eatables.ts
@@ -1,4 +1,8 @@
-import { Eatable } from '../skilling/types';
+export interface Eatable {
+	name: string;
+	id: number;
+	healAmount: number;
+}
 import itemID from '../util/itemID';
 
 export const Eatables: Eatable[] = [
